Change type to complex64 for benchmarks

+++ src/qibo/config.py
@@ -21,9 +21,9 @@ if BACKEND_NAME == "tensorflow":
     import tensorflow as tf
 
     # Default types
-    DTYPE = tf.float64
+    DTYPE = tf.float32
     DTYPEINT = tf.int32
-    DTYPECPX = tf.complex128
+    DTYPECPX = tf.complex64
 
     from qibo.tensorflow import gates
     from qibo.tensorflow import circuit as tensorflow_circuit
